- added byte instead of float - removed unitaction history - tried GA with bitfields -...

+++ ETF_EVOL_MODEL/src/org/eft/evol/stats/UnitAction.java
@@ -8,15 +8,15 @@ public class UnitAction {
 	public int actionType;
 	public int cycle;
 	public int iteration;
-	public long shares;
+	public int shares;
 	public float nav;
 	public int indexOfETF;
 	public float gradient;
-	
+
 	@Override
 	public String toString() {
 		StringBuilder builder = new StringBuilder();
-		
+
 		builder.append(iteration);
 		builder.append(",");
 		builder.append(cycle);
@@ -25,16 +25,14 @@ public class UnitAction {
 		builder.append(",");
 		builder.append(shares);
 		builder.append(",");
-		builder.append((actionType == ActionType.HOLD) ? "":ETFMap.getInstance().getEtfName(indexOfETF));
+		builder.append((actionType == ActionType.HOLD) ? "" : ETFMap.getInstance().getEtfName(indexOfETF));
 		builder.append(",");
 		builder.append(nav);
 		builder.append(",");
 		builder.append(gradient);
 		builder.append("\n");
-		
+
 		return builder.toString();
 	}
-	
 
-	
 }
